# openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/webhooks":{"get":{"tags":["webhooks-controller"],"summary":"Lists all registered webhooks","description":"Returns all webhook endpoints registered in the system.","operationId":"findAll","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"List of webhooks returned successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Webhook"}}}}}},"post":{"tags":["webhooks-controller"],"summary":"Create a new webhook","description":"Creates a new webhook record. This webhook will be notified whenever a new payment is created.","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Webhook"},"examples":{"Webhook example":{"description":"Webhook example","value":{"url":"http://localhost:8080/api-mock/success"}}}}},"required":true},"responses":{"400":{"description":"Invalid webhook payload","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Webhook created successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/payments":{"get":{"tags":["payments-controller"],"summary":"Lists all payments","operationId":"list","responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Payment"}}}}}}},"post":{"tags":["payments-controller"],"summary":"Create a new payment","description":"Creates a payment record in the system and notifies all existing webhooks with the payment content.","operationId":"save_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payment"},"examples":{"Payment example":{"description":"Payment example","value":{"firstName":"John","lastName":"Doe","zipCode":"12345","cardNumber":"4111111111111111"}}}}},"required":true},"responses":{"400":{"description":"Invalid request","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Payment created successfully","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api-mock/success":{"post":{"tags":["api-mock-controller"],"summary":"Mock a successful response","description":"Mocks a response with HTTP status 200 (OK). This endpoint logs the provided payload and returns a status 200 response. Useful for testing purposes.","operationId":"success","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api-mock/fail":{"post":{"tags":["api-mock-controller"],"summary":"Mock a failure response","description":"Mocks a response with HTTP status 400 (Bad Request). This endpoint logs the provided payload and returns a status 400 response. Useful for testing failure scenarios.","operationId":"fail","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/webhooks/{id}":{"delete":{"tags":["webhooks-controller"],"summary":"Remove a webhook","description":"Deletes a webhook by its unique identifier.","operationId":"deleteById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"204":{"description":"Webhook deleted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Webhook not found","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Webhook":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"callbackUrl":{"type":"string"}}},"Payment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"zipCode":{"type":"string"},"cardNumber":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}}}}